fix(media): normalize abort rejection to MediaFetchError during retry backoff
ClawSweeper P1: sleepWithAbort rejects raw Error("aborted") when the
caller signal fires during backoff. retryAsync awaits the sleep outside
the attempt callback, so the raw error bypassed the media error
normalization in fetchGuardedMediaResponse. This broke the established
contract where all abort paths surface MediaFetchError with code
"fetch_failed" and an AbortError cause.

Catch the sleepWithAbort rejection in the custom sleep wrapper and
convert it to MediaFetchError before it propagates through retryAsync.

Also extract content-disposition parsing helpers to
src/media/content-disposition.ts to keep fetch.ts within the 700-line
max-lines budget after adding the error normalization.

Update the regression test to assert the full MediaFetchError shape
(name, code, cause.name) instead of just toThrow().
