+++ src/media/fetch.ts
@@ -4,7 +4,7 @@ import { parseMediaContentLength } from "@openclaw/media-core/content-length";
 import { basenameFromAnyPath, extnameFromAnyPath } from "@openclaw/media-core/file-name";
 import { detectMime, extensionForMime } from "@openclaw/media-core/mime";
 import { expectDefined } from "@openclaw/normalization-core";
-import { isAbortError } from "../infra/abort-signal.js";
+import { createAbortError, isAbortError } from "../infra/abort-signal.js";
 import { sleepWithAbort } from "../infra/backoff.js";
 import { formatErrorMessage } from "../infra/errors.js";
 import {
@@ -22,6 +22,7 @@ import { retryAsync, type RetryOptions } from "../infra/retry.js";
 import { isTransientNetworkError } from "../infra/retryable-network-errors.js";
 import { redactSensitiveText } from "../logging/redact.js";
 import { buildTimeoutAbortSignal } from "../utils/fetch-timeout.js";
+import { basenameFromUrlPathname, parseContentDispositionFileName } from "./content-disposition.js";
 import { saveMediaBuffer, saveMediaStream, type SavedMedia } from "./store.js";
 
 /** Default remote media fetch cap shared by buffer reads and store writes. */
@@ -129,135 +130,6 @@ type GuardedMediaResponse = {
   release: (() => Promise<void>) | null;
   sourceUrl: string;
 };
-
-function stripQuotes(value: string): string {
-  return value.replace(/^["']|["']$/g, "");
-}
-
-function decodeRemoteFileNameComponent(value: string): string {
-  try {
-    return decodeURIComponent(value).replace(/[\\/]/g, "_");
-  } catch {
-    return value;
-  }
-}
-
-function decodeExtendedRemoteFileName(value: string): string | undefined {
-  const match = /^([^']*)'[^']*'(.*)$/u.exec(value);
-  if (!match) {
-    return undefined;
-  }
-  const charset = match[1]?.toLowerCase();
-  const encoded = match[2] ?? "";
-  try {
-    if (charset === "utf-8") {
-      return decodeURIComponent(encoded).replace(/[\\/]/g, "_");
-    }
-    if (charset === "iso-8859-1") {
-      if (/%(?![\da-f]{2})/iu.test(encoded)) {
-        return undefined;
-      }
-      return encoded
-        .replace(/%([\da-f]{2})/giu, (_match, hex: string) =>
-          String.fromCharCode(Number.parseInt(hex, 16)),
-        )
-        .replace(/[\\/]/g, "_");
-    }
-  } catch {
-    return undefined;
-  }
-  return undefined;
-}
-
-function* parseContentDispositionParameters(header: string): Generator<{
-  name: string;
-  value: string;
-}> {
-  let start = 0;
-  let quoted = false;
-  let escaped = false;
-  for (let index = 0; index <= header.length; index += 1) {
-    const character = header[index];
-    if (escaped || (quoted && character === "\\")) {
-      escaped = !escaped;
-      continue;
-    }
-    if (character === '"') {
-      quoted = !quoted;
-      continue;
-    }
-    if (index !== header.length && (quoted || character !== ";")) {
-      continue;
-    }
-    const parameter = header.slice(start, index).trim();
-    start = index + 1;
-    const separator = parameter.indexOf("=");
-    if (separator > 0) {
-      yield {
-        name: parameter.slice(0, separator).trim().toLowerCase(),
-        value: stripQuotes(parameter.slice(separator + 1).trim()),
-      };
-    }
-  }
-}
-
-function decodeQuotedRemoteFileName(value: string): string {
-  const windowsDrivePath = /^[a-z]:[\\/]/iu.test(value);
-  const windowsNetworkPath = value.startsWith("\\\\");
-  const mixedWindowsPath = value.includes("/") && value.includes("\\");
-  const relativeWindowsPath =
-    /\\[\p{L}\p{N}]/u.test(value) && /^[^\\/:]+(?:\\[^\\]+)+$/u.test(value);
-  if (!windowsDrivePath && !windowsNetworkPath && !mixedWindowsPath && !relativeWindowsPath) {
-    return value.replace(/\\(.)/gu, "$1");
-  }
-  const lastForwardSeparator = value.lastIndexOf("/");
-  if (lastForwardSeparator >= 0) {
-    const prefix = value.slice(0, lastForwardSeparator + 1);
-    const fileName = value.slice(lastForwardSeparator + 1).replace(/\\([^\p{L}\p{N}])/gu, "$1");
-    return `${prefix}${fileName}`;
-  }
-  const firstBackslash = value.indexOf("\\");
-  if (
-    !windowsDrivePath &&
-    !windowsNetworkPath &&
-    firstBackslash === value.lastIndexOf("\\") &&
-    /\\[^\p{L}\p{N}]/u.test(value)
-  ) {
-    return value.replace(/\\(.)/gu, "$1");
-  }
-  // Backslash-only legacy paths need every separator, including before Unicode or spaces.
-  return value.replace(/\\"/gu, '"');
-}
-
-function parseContentDispositionFileName(header?: string | null): string | undefined {
-  if (!header) {
-    return undefined;
-  }
-  let fallbackFileName: string | undefined;
-  for (const parameter of parseContentDispositionParameters(header)) {
-    if (parameter.name === "filename") {
-      fallbackFileName ??=
-        basenameFromAnyPath(decodeQuotedRemoteFileName(parameter.value)) || undefined;
-      continue;
-    }
-    if (parameter.name !== "filename*") {
-      continue;
-    }
-    const decoded = decodeExtendedRemoteFileName(parameter.value);
-    if (decoded) {
-      return basenameFromAnyPath(decoded) || undefined;
-    }
-  }
-  return fallbackFileName;
-}
-
-function basenameFromUrlPathname(pathname: string): string {
-  const base = basenameFromAnyPath(pathname);
-  if (!base) {
-    return "";
-  }
-  return decodeRemoteFileNameComponent(base);
-}
 
 async function readErrorBodySnippet(
   res: Response,
@@ -638,12 +510,28 @@ async function withMediaFetchRetry<T>(
   if (!retry) {
     return await fn();
   }
+  // sleepWithAbort rejects raw Error("aborted") on caller signal abort during
+  // backoff; normalize to MediaFetchError so callers retain the established
+  // fetch_failed / AbortError-cause contract used by header and body aborts.
+  const signal = options.requestInit?.signal ?? undefined;
+  const sleep =
+    retry.sleep ??
+    ((ms: number) =>
+      sleepWithAbort(ms, signal).catch((err) => {
+        throw new MediaFetchError(
+          "fetch_failed",
+          `Failed to fetch media from ${redactMediaUrl(options.url)}: ${formatErrorMessage(err)}`,
+          {
+            cause: createAbortError("aborted", { cause: err }),
+          },
+        );
+      }));
   return await retryAsync(fn, {
     label: "media:fetch",
     ...retry,
     shouldRetry: (err, attempt) =>
       retry.shouldRetry ? retry.shouldRetry(err, attempt) : shouldRetryMediaFetch(err),
-    sleep: retry.sleep ?? ((ms) => sleepWithAbort(ms, options.requestInit?.signal ?? undefined)),
+    sleep,
   });
 }
